youtube-cleanup: remove Hashtag above video title
The existing rules don't work with the new youtube layout (not another new one, but the current new one). *Some how never tested this rule before after the change.* 

If you play a video with hashtags - like `https://www.youtube.com/watch?v=llxmTHt5zPs (guitar music)` the hashtags are not shown above the video title, but at the end of the line which shows the date the video got uploaded. 

See: