google-search-cleanup: Hides "People also search for" on mobile versions
Link used for testing: `https://www.google.com/search?q=remove+people+also+search+for#ip=1`

The rules for `People also ask` also works on mobile versions, but `People also search for` not. 
The proposed rule did not give false positives both on the mobile version as on the normal version of google.